--- react-ui/src/components/index.ts
@@ -1,4 +1,0 @@
-export * from "./Lockscreen";
-export * from "./UnlockModal"
-export * from './GatedRoute';
-export { TokiumContext, TokiumProvider } from './TokiumContext';
