+++ react-ui/src/index.css
@@ -33,7 +33,7 @@ code {
   flex-direction: row;
   flex-wrap: wrap;
   flex-shrink: 0;
-  align-items: start;
+  align-items: flex-start;
   margin-top: 5em;
   row-gap: 5em;
   column-gap: 5em;
