+++ react-ui/src/wrapper/components/gatedRoute/index.ts
@@ -1,0 +1,1 @@
+export * from './GatedRoute';
