+++ react-ui/src/wrapper/components/unlockModal/index.ts
@@ -1,0 +1,1 @@
+export * from './UnlockModal';
