+++ react-ui/src/wrapper/index.ts
@@ -1,0 +1,4 @@
+export * from "./components/lockScreen/Lockscreen";
+export * from "./components/unlockModal/UnlockModal"
+export * from './components/gatedRoute';
+export { TokiumContext, TokiumProvider } from './context';
